requirements.txt
IPython.display
demo
flask
imageio
imageio_ffmpeg
matplotlib.animation
matplotlib.pyplot
numpy
os
pyngrok
skimage
skimage.transform
tempfile
torch
warnings